Pages security

+++ app/Http/Controllers/ApartmentController.php
@@ -2,48 +2,26 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StoreApartmentsRequest;
 use App\Models\Apartment;
 use App\Models\Booking;
 use App\Models\Image;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 
 class ApartmentController extends Controller
 {
-    public function create(Request $request)
+    public function create(StoreApartmentsRequest $request)
     {
-        $request->validate([
-            'title' => ['required', 'max:255'],
-            'description' => ['required', 'max:300'],
-            'rooms' => ['required', 'integer'],
-            'max_people' => ['required', 'integer'],
-            'price' => ['required', 'integer'],
-            'photos' => ['required', 'array'],
-            'photos.*' => ['image', 'mimes:jpeg,png,jpg', 'max:10240'],
-            'longitude' => ['required'],
-            'latitude' => ['required'],
-            'address' => ['required'],
-        ]);
-
-        $address = $request->input('address');
-        $address = str_replace(', ', ',', $address);
-        $address = explode(',', $address);
-        $city = explode(' ', $address[1]);
+        $address = explode(', ', $request->address);
 
         // Create an apartment
-        $apartment = new Apartment([
-            'title' => $request->input('title'),
-            'description' => $request->input('description'),
-            'max_people' => $request->input('max_people'),
-            'rooms' => $request->input('rooms'),
-            'price' => $request->input('price'),
-            'lon' => $request->input('longitude'),
-            'lat' => $request->input('latitude'),
-        ]);
+        $apartment = new Apartment($request->validated());
         $apartment->country = $address[2];
-        $apartment->city = $city[1];
+        $apartment->city = $address[1];
         $apartment->street = $address[0];
 
         $apartment->owner()->associate(Auth::user());
@@ -63,63 +41,40 @@ class ApartmentController extends Controller
 
     public function detail($id)
     {
-        $chosenApartment = Apartment::where('id',$id)->with('images')->firstOrFail();
+        $chosenApartment = Apartment::getApartment($id);
         $existedBookings = Booking::where('apartment_id', $id)->get();
         return view('details', ['apartment' => $chosenApartment, 'bookings' => $existedBookings]);
     }
 
     public function delete($id)
     {
+        Gate::authorize('user_is_owner', Apartment::getApartment($id));
         $containedImages = Image::where('apartment_id', $id)->get();
         foreach($containedImages as $image)
         {
             Storage::disk('public')->delete('images/' . $image->path);
             $image->delete();
         }
-        $apartment = Apartment::where('id', $id)->delete();
+        $apartment = Apartment::getApartment($id)->delete();
         return redirect('/');
     }
 
     public function edit_index( $id)
     {
-        $editaibleApartment = Apartment::where('id',$id)->with('images')->first();
-
-        return view('edit', ['editApartment' => $editaibleApartment]);
+        $editableApartment = Apartment::getApartment($id);
+        Gate::authorize('user_is_owner', $editableApartment);
+        return view('edit', ['editApartment' => $editableApartment]);
     }
 
-    public function edit(Request $request, $id)
+    public function edit(StoreApartmentsRequest $request, $id)
     {
-        $request->validate([
-            'title' => ['required', 'max:255'],
-            'description' => ['required', 'max:300'],
-            'rooms' => ['required', 'integer'],
-            'max_people' => ['required', 'integer'],
-            'price' => ['required', 'integer'],
-            'photos' => ['array'],
-            'photos.*' => ['image', 'mimes:jpeg,png,jpg', 'max:10240'],
-            'longitude' => [''],
-            'latitude' => [''],
-            'address' => ['required'],
-        ]);
+        $apartToEdit = Apartment::getApartment($id);
+        Gate::authorize('user_is_owner', $apartToEdit);
+        $address = explode(', ', $request->address);
 
-        $apartToEdit = Apartment::where('id', $id)->with('images')->first();
-
-        $address = $request->input('address');
-        $address = str_replace(', ', ',', $address);
-        $address = explode(',', $address);
-        $city = explode(' ', $address[1]);
-
-
-        $apartToEdit->title =  $request->input('title');
-        $apartToEdit->description = $request->input('description');
-        $apartToEdit->max_people = $request->input('max_people');
-        $apartToEdit->rooms = $request->input('rooms');
-        $apartToEdit->price = $request->input('price');
-        $apartToEdit->lon = $request->input('longitude');
-        $apartToEdit->lat = $request->input('latitude');
-
+        $apartToEdit->fill($request->validated());
         $apartToEdit->country = $address[2];
-        $apartToEdit->city = $city[1];
+        $apartToEdit->city = $address[1];
         $apartToEdit->street = $address[0];
 
         $apartToEdit->save();
